pyproject.toml
[project]
name = "django-ninja-mcp"
version = "0.1.0"
description = "MCP extension for django-ninja"
readme = "README.md"
requires = ["Django >=3.1", "pydantic >=2.0,<3.0.0", "django-ninja > 1.0"]
description-file = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-ninja>=1.4.1",
    "mcp>=1.6.0",
    "pydantic>=2.5.3",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-asyncio",
    "psycopg2-binary",
    "mypy==1.7.1",
    "ruff==0.5.7",
    "django-stubs",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"
exclude = ["store/clients","docs","athena_models"]

[tool.ruff.lint]
select = ["F", "E", "W", "I001", "D"]
ignore = ["D203", "D212", "D10"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 120